fix(db): invalidate caches after async pet writers
