+++ src/app/api/admin/edits/[id]/route.ts
@@ -8,12 +8,12 @@ import { isAdmin } from "@/lib/admin";
 import {
   AGGREGATE_KEYS,
   invalidateAggregates,
+  invalidatePetCaches,
 } from "@/lib/db/cached-aggregates";
 import { db, schema } from "@/lib/db/client";
 import { renderEditApprovedEmail } from "@/lib/email-templates/edit-approved";
 import { renderEditRejectedEmail } from "@/lib/email-templates/edit-rejected";
 import { createNotification } from "@/lib/notifications";
-import { invalidatePetCaches } from "@/lib/pets";
 import { deleteR2Objects, keyFromR2Url } from "@/lib/r2";
 import { requireSameOrigin } from "@/lib/same-origin";
 import { refreshSimilarityFor } from "@/lib/similarity";
